xfs_copy: copy/restore xfs filesystem to/from regular file
When copying to regular file, the length of the resulting file
is equal to the size of the source filesystem. I added an option
which store only the amount of space actually used  in the source
filesystem by the filesystem and the XFS log.

$ xfs_copy -s /dev/sdd1 regularfile
copy xfs filesystem on /dev/sdd1 to regular file
as the amount used in fs.

$ xfs_copy -f regularfile /dev/sdd1
copy xfs filesystem to /dev/sdd1 from regular file
which was copied using "xfs_copy -s"